Issue #7 - Make UI dependency installation optional
The custom `install` and `develop` commands have been expanded to both
take an optional installation parameter called 'with-ui-deps' that
controls whether the UI dependencies will be installed or not. By
default, this is set to False. The command definitions have been split
as well to facilitate future changes as part of the build such as
automatic re-building of the bundled files for `develop`.

You can install the dependencies with
python setup.py [install|develop] --with-ui-deps=True

If you're using pip you will need to wrap the options
pip install [-e] . --install-option="--with-ui-deps=True"