Issue #9 - Update bliss.gui package structure
bliss.gui is now a subpackage with the relevant static files for the gui
placed inside of 'static'. The existing bliss.gui module has been
temporarily moved to bliss.gui.gui. This will need to be reworked to
reduce redundancy and make the module split more intelligible. Follow up
tickets will cover that.