Remove util functions that duplicate existing Lodash functionality
Lodash is now required in index.js. Note, at the moment all of Lodash is
being included. We can update this to include just the functions we care
about if we're concerned with the size. The bliss.util map, mapN, and
range functions have been replaced with the Lodash functions that
provide the same functionality.