Update Messages with scrollable display
Message are now displayed in a scrollable table at the bottom of each
tab. The implementation for this was moved from a table to bootstrap
container/row/columns due to issues styling the scrolling table. The
message display has been updated to include the message severity. There
was a bug with the access of the attribute for the value in the message
so it was missing from the display.