Issue #29 - Add support for bliss-core install from PyPi
- Update README documentation with info on how to specify additional
PyPi servers so bliss-core can be installed as part of the GUI
installation. The developer install instructions had to be reworked due
to issues with pip passing install-option's to dependencies as well as
the package to be installed. The current supported work around is a
horrible hack. See https://github.com/pypa/pip/issues/1883 for info.
- Updated the bliss-core version to the latest release.