Issues #95 and #96 - Fix static file routing problems
Remove HTMLRoot.Fallback in favor of a less specific Static route
definition that excludes the 'build' directory. This is in-line with
the previous assumptions about static file routing as found in
index.html. This routing "matches" our directory layout between
index.html and the repo. E.g., the static file `bliss.bundle.js` path
from index.html is a valid path to the file from the bliss-gui
repository root.

HTMLRoot.User now defaults to the Static path defined above. This means
that a clean checkout of bliss-gui and bliss-core now work out of the
box.