Issue #120 - Pull Sequence and command history paths from config
The Sequence file directory setting is now pulled from bliss.config with
the `sequence.directory` attribute. If that directory doesn't exist it
defaults to None and is logged as a warning.

The Command History file path setting is now pulled from bliss.config
with the `command.history.filename` attribute. If the file doesn't
currently exist and the dirname for the file points to a directory that
doesn't exist, a warning is logged informing the user that command
history will not be saved.