Issue #118 - Misc. Script component cleanup and polishing
Script execution state is now tracked in a global object instead of in
the Script component. This resolves a number of headaches associated
with the Script component (and all its child components) getting removed
when changing tabs and being reinstated later with bogus default state.

Fixes CodeMirror rendering handling in the ScriptEditor component after
tab changes. The view method was being called after oninit and before
oncreate after a tab change and this was resulting in invalid variable
states.

The editor window now scrolls when the currently executing line rolls
off the screen. Note, the window has to be scrolled to the pre-update
location each update to ensure that we don't end up with flickering /
the window being scrolled to the top each update regardless of where the
user had scrolled it manually.

Accidentally fixed a "select flickering" issue by adding a comment
before the `bliss-scripts` component tag in index.html. Without this the
multi-select script box will constantly scroll to the top of the options
list with each redraw cycle. Not sure why it's doing this and even less
sure why the comment fixes it ...