Issue #169 and #146 - Component DOM access and event handler cleanup
Component's event handlers are now all defined in the attributes that
are passed to `m()` instead of in `oncreate`. This removes the need to
have id's reference specific items in a number of cases.

Script components that were directly referencing DOM elements via id now
reference them via `vnode.dom`. This eliminates the need for specific
ids on some components.