Issue #156 - Add default format for decimal places
If a Field does not have a format specifier and is displaying a floating
point value we default to five decimal places in the display. This should
resolved overly large floating point numbers overlapping in the display
and gives Field a sane default for floats if the user forgets to specify
a format for it.