Issue #173 - Event handlers should use e.currentTarget for DOM access
Event handlers in Sript/Command components have been updated to use
e.currentTarget for access the DOM element to which the event handler
was attached instead of e.target. This ensures that the correct base
element is used when referencing elements in relation to the target
element.

This is a follow up to #169 and #146