Issue #56 - Make entire command accordion header click-able
Headers for accordions in the command browser are now entirely clickable
instead of requiring the user to click the category text. The structure
of the HTML output by the command components has been updated to match
the expected structure for bootstrap accordions with regards to panel
divs surrounding each header/body pair. The rendering before was
incorrectly nesting all of the header/body pairs inside of a single
"panel" div.

Accordion behavior has been changed slightly by removing data-parent
references. Accordions that have been opened now stay open even if
another accordion is opened. This makes more sense from a data
exploration perspective and addresses an issue with accordions closing
sometimes when `collapse('show')` is called. There were a number of
bootstrap tickets referencing this or similar problems, but evidently
the problem still persists in the version we're using.