Issue #203 - Better handle unknown EVR/CMD display
EVR and Cmd dictionaries `getByCode` functions now return the relevant
dictionary entry or the supplied code if no entry exists for it. This
means that raw values will be shown when an EVR/Cmd dictionary doesn't
match the telemetry instead of 'N/A'. Note, the 'N/A' message was being
shown because the value for the field was being set as undefined due to
the indexing into the dictionaries with invalid values.

EVR/Cmd display in the Field component is now handled more clearly. In
practice these changes shouldn't make a difference and the updated
"error" messages should never be seen but if we do it'll be more
obvious compared to the previous situation where 'N/A' was displayed in
multiple situations