Issue #208 - Remove script load modal from UI
The script load modal has been removed in favor of a toggle-able div
present in the Scripts tab.

Increased the width of Script components so they take up the entire 12
columns.

The ScriptExecCtrl component now takes a mandatory attribute called
loadButtonAction that is called when the load button is clicked. This
allows parent components to control how they want loading handled. For
instance, the load button could toggle the ScriptLoadModal component
display as it was doing before by passes a function which toggles it.

Added an explicit refresh of the CodeMirror element in the ScriptEditor
view method. This addresses some issues with improper rendering of
the element such as missing gutters and text overlapping the gutters.