Issue #214 - Remove bootstrap grid from Messages component
The Messages component now uses flexboxes for display of messages
instead of rows in the Bootstrap grid. This gives us more flexibility
with the scaling of the Messages display compared to the grid approach.
If we decide to scale the UI to phone/tablet sizes this will give us a
solid baseline for adjustments.

Specific ids have been removed from the component so multiple
bliss-messages components can exist on the same page without issue.

CSS has been moved to a component specific file and updated with a
(hopefully) cleaner naming scheme.

Scrolling has been updated so that the most recent message is at the
bottom of the scroll window. If a user scrolls back through the history
their location is saved when the component is re-rendered. Scrolling
back to the most recent item resets the scrolling history. Would be
useful to consider adding a button to reset the scroll in the future.