Issue #235 - Support Basic HTTP Strict Transport Security (HSTS)
When the BLISS GUI is installed behind an Apache or Nginx SSL web
proxy (e.g. for JPL LDAP authentication), all protocol URLs must use
SSL between the client and the proxy to prevent protocol downgrade
attacks.  This is known as [HTTP Strict Transport Security
(HSTS)](https://en.wikipedia.org/wiki/HTTP_Strict_Transport_Security)
and is the default in all modern browsers.

This isn't a problem for most URLs in the GUI as they're never
explicit about their protocol, so when loaded over `https://`, for
example, all URL references are `https://`.  The one exception is the
websocket connection in `bliss/gui/index.js` for realtime telemetry
display, which requires an explicit `ws://` or `wss://`.  BLISS GUI
should choose whether or not to connect to the websocket using SSL
based on how the page was loaded, i.e.:

    const proto = location.protocol === 'https:' ? 'wss' : 'ws'
    const url   = proto + '://' + location.host + '/tlm/realtime'