Issue #244 - Add support for aborting running scripts
Add support for aborting a running script via the /script/abort/
endpoint. The script execution control panel has been updated to include
an abort button that allows scripts that are running / paused to be
aborted. Upon script abortion, the execution state is reset to the
initial state after loading.

Dropped the "reset" button placeholder from the execution control panel.
The abort button functionality overlaps sufficiently with reset that
it's unlikely we'll implement reset in the future.