Issue #253 - Clean up Commanding view layout
Fix perpetual horizontal scroll bar in bliss-commandconfig display.
Update the bliss-commandsearch component to better handle accordion
hiding / display based on whether a command is currently selected.

If no command is selected for config:
    - Entering the search area will expand all the accordions
    - Leaving the search area will not close the accordions

If a command is selected for config:
    - Entering the search area will not expand all the accordions
    - Leaving the search area will expand all the accordions

In both cases, the search box maintains its previous functionality of
expanding all accordions on focus.