Issue #269 - Clean up Command components css
Command component CSS has been moved to a Command specific module and
updated to BEM. All `id`s have been removed from the command components
and the code has been updated as needed to handle this.

The command browser section has been updated to make better use of the
screen space provided by the horizontal layout.

Minor change to the way the command browser accordions auto open/close
so they will be shown if the user hovers over after having entered some
search terms regardless of whether or not a command has been selected.

Fixed bug with the CSS grid layout handling where the `cmdbrowser` grid
area was being erroneously set in both the cmd-search and cmd-browser
containers. Browsers seemed to handle this intelligently, so it didn't
manifest in a noticeable way, but it was incorrect nonetheless.