Issue #295 - Add support for CCSDS telemetry streams
This commit adds the capability to have multiple packet types arrive
on the same port, differentiated by their CCSDS header.  In that case,
a packet `name` isn't required and `type` (which defaults to `raw` if
not specified) should be `ccsds`, e.g.:

    gui:
        telemetry:
            - stream:
                port: 3076
                type: ccsds

This commit changes `UdpTelemetryServer` to `UdpRawTelemetryServer`
and adds a UDP telemetry server for CCSDS packets, so we now have:

  - UdpCcsdsTelemetryServer
  - UdpRawTelemetryServer

It also adds a signficant amount of error checking when configuring
GUI packet streams.

APIDs of zero are now supported in the GUI.

The default GUI syslog port (2514) can now be changed via config.yaml:

    logging:
        port: 4242

Fixes #295.