Issue #310 - Sequence send button is not responding
The check for sequence form submission with no sequence file selected
was formatted incorrectly and checking for an incorrect return value.
This was causing all the form submissions to get blocked and no
sequences run.

The Enter key has been disabled on the entire sequence select form to
prevent users from accidentally submitting the form.