Issue #333 - Add command validation to CommandInput
Add command validation functionality to the CommandInput component.

Update CommandInput view to render `Send` button based on state of
command validation. If a valid command has not been entered the send
button will be red and disabled. If validation is occurring the button
will be a red, disabled, and show a spinning refresh icon. If a valid
command has been entered the button will be green and enabled.

Submission handling has been updated to prevent submission of commands
that don't pass validation.

Update the backend with the `/cmd/verify` endpoint for validation of
commands. The endpoint expects to receive the command for verification
in the same format as `/cmd`. Namely, form data where the command string
is stored under the `command` key.

Command creation and validation code has been factored out of `/cmd`
into a helper function (_create_and_verify_command) so it can be used by
both `/cmd` and `/cmd/verify`.