Issue #356 - Allow user customization of Modal component properties
Allow customization of Modal component parameters via configuration
passed in the body of the toggling modal:show event.

The following attributes were added:
    displayBackground: Toggle display of the modal background. Defaults
        to true.

    insertHeaderCloseBtn: Specifies if a close button should be inserted
        into the modal header for the user. Defaults to true

    insertFooterCloseBtn: Specifies if a close button should be inserted
        into the modal footer for the user. Defaults to true

The expected data format for the modal header contents has changed. Now
the Modal component expects to receive the contents of the modal title
in the 'header' attribute of the event body. The component handles
wrapping the title in a properly classed h4 tag for the user. The code
affected by this has been updated to work with the new requirements.