Issue #354 - Add logging of tlm snapshot value in search results modal
Add logging of the telemetry field's value at time of inspection in the
mnemonic search result modal. Note, this is currently extracting values
for logging by inspecting the current packet at the time of result
viewing. This will need to be updated when the Component rendering
problems are fixed so that the values are updating in real time.

Add backend POST /messages endpoint for submitting a message for
logging. The end point expects to receive a severity and the message to
log. If it is unable to locate an appropriate handler for the requested
severity a default of `info` is used.