Issue #375 - Improve Highcharts plotting performance
Update the plotting routine in the HighchartsBackend object so that
points are added to the plot with redraw disabled and a single redraw()
call is made after all plot updates have occurred.

Tweak Highcharts settings to (hopefully) increase performance.
`boostThreshold`, `cropThreshold`, and `turboThreshold` have all been
set to 1. Options for disabling animations, mousetracking, and other
misc. niceties have been added to the configuration objects, documented,
and commented out in case we choose to use them later.