Issues #379 and #380 - Clean up tlm popover generation and add disable
Clean up the Field telemetry popover content generation code with
template literals and better tag choices. Styling has been moved
to Field.css.

Add an optional attribute to `bliss-field` for disabling popovers on
fields. If `disable-tlm-popover` is present on the `bliss-field` tag
then the telemetry popover will not be displayed.