Issue #391: Update Dygraphs Options and Sliding Window
This update includes:

* bug fix for the missing starting array
* use bliss-plotconfig to set necessary dygraphs attributes

The 'labelsDiv' is the most critical to make the legend look
reasonable.

Also have some logic to pop values off of the plot.data to avoid
crashing the browser with too many data points.

Here is a quick example I created for ECOSTRESS. The <table> helps
with some of the weird formatting we were seeing. Definitely
open for other suggestions or improvements.

<table align="center">
  <tr>
    <td>
      <bliss-plot plot-range="600">
        <bliss-plot-config>
          {
            "ylabel":     "Kelvin",
            "height":     500,
            "width":      1000,
            "labelsDiv":  "cryo-3-plot-legend"
          }
        </bliss-plot-config>
        <bliss-plot-series packet="ECO_1553_EHS" name="Cryo3TempAmb" caption="TempAmb"></bliss-plot-series>
    </td>
    <td>
      <div id="cryo-3-plot-legend">
      </div>
    </td>
  </tr>
</table>