Issue #401 - Clean up Dygraphs plot legends
Dygraph plots are drawn with an external div legend by default. The
legend formatting has been adjusted so the legend is displayed under the
plot with labels inline.

DygraphBackend initialization has been changed to handle the external
div legends. First, `bliss-plot` tag rendering has been updated when
Dygraphs is being used to include 2 nested `div`s within `bliss-plot`.
The first is used to initialize Dygraphs and the second is used for the
legend. This is required so that Dygraphs isn't initialized on the
legends parent element (as we were doing before by using `vnode.dom`).
The width of the legend `div` is injected into the element creation in
`view` so that we can style it to match the corresponding plot. Each
instance of DygraphBackend uses a Math.random() float to uniquely
identify it and its' corresponding label `div`. This number is used as
the id for the legend and should be unique enough to avoid clashes on
the page.