Issue 435 - DN to EU conversion fails with multiple packet types
Defining more than one type of `!Packet` may cause client-side DN to EU
conversions to fail on all but the first packet received by the client.
This is because the Javascript `bliss.tlm.Packet` constructor guards
against the creation of packet properties more than once:

    if (Packet.prototype.__init__ === undefined) {
        for (const name in this._defn.fields) {
            //const getter = () => this.__get__(name)
            Object.defineProperty(Packet.prototype, name, {
                get: function () {
                    return this.__get__(name)
                }
            })
        }
        Packet.prototype.__init__ = true
    }

If a packet property (i.e. `!Field`) is referenced in a DN to EU equation,
but is not present on the packet object being evaluated, Javascript will
return `undefined` which will result in either an `undefined` or `NaN`
being returned for the DN to EU conversion.

This bug is easily reproduced by starting the BLISS MOXIE GUI and sending
two zeroed packets form the Python interpreter:

    >>> import socket
    >>> sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
    >>> sock.sendto(bytearray(62) , ('127.0.0.1', 11111))  # MAIN_HK
    >>> sock.sendto(bytearray(141), ('127.0.0.1', 11115))  # STD_TLM_PKT

For whichever packet is sent second, the DN to EU conversion will fail.

The cause of the bug (@bornstei) is remarkably shortsighted, allowing only
one packet type (with DN to EU conversions). The fix is to *properly* allow
multiple packet types to be defined in the GUI by creating a subclass for
each (at runtime).

Thanks to @dominick for narrowing down exactly how to reproduce this issue.
cc: @mjjoyce

Fixes #435