Issue #439 - Update Plots with configurable redraw frequency
Update the Plot component to support configurable redraw frequency. Plot
redraw frequency can be set in seconds with the `redraw-frequency` attribute
on the `bliss-plot` tag. The plot will refresh once `redraw-frequency`
seconds have passed since the last redraw attempt. By default, the
`redraw-frequency` is set to 10 seconds.

Support for adding random variation to the redraw frequency can be
toggled and configured with the `redraw-frequency-variation`,
`redraw-frequency-variation-min`, and `redraw-frequency-variation-max`
attributes. If `redraw-frequency-variation` is set to true, the redraw
frequency will be adjust by a random float in the range
[`redraw-frequency-variation-min`, `redraw-frequency-variation-max`).
By default, the random adjustments are turned off and the value range is
to [-2, 2)