Issue #443 - Fix Dygraph Backend multi-packet plot handling
Update Dygraph backend to properly format data added to the plot when
handling multiple packets in a single plot. The backend now tracks the
ids (packet.field) for each series of data in the plot. When redrawing
due to a packet receipt, the backend uses this array of series ids to
update fields that are contained in the packet and also insert null
values for fields that are in other packets. This ensures that the data
appended to the plot maintains the proper "shape".