Issue #25 - Update Plot time label and Search time handling (#30)
Update the plot x-axis label to `Time` instead of `Time (UTC)`. This
field isn't necessarily UTC since users can specify a custom field in
their telemetry that indicates time. The more generic label should help
avoid confusion here.

Update time handling in MnemonicSearch to explicitly use UTC time
instead of the default (which was changed in #3). In the future we
should attempt to update the handling of this so that the Clock mode
dictates which time format is used.