Issue #82 - Add commands without subsystem to default subsystem group
Commands are added to a default subsystem group on the frontend
during command dictionary parsing so that components that expect
commands to be grouped as such function as expected. Specifically, this
allows the CommandSearch object to get commands grouped as expected for
display. Since the subsystem attribute isn't required this means that the
CommandSearch component will default to all commands showing up under
this default subsystem header instead of not at all.