Issue #85 - Add support for EVR message formatting
Update EVRDefinition to support optional 'message' attribute that is
currently supported in Core. The message attribute is a string that can
contain printf format strings that can be processed via
EVRDefinition.formatMessage().

Add EVRDefinition.formatMessage() for formatting an EVR's message
attribute with an array of data values. Most commonly this would be used
to format an EVR's info message with info returned in H&S data to
provide additional information to operators.

Fix typo in Field oninit where this._cached.rawval was being defaulted
to null instead of this._cached.val which is used throughout the code.

Update Field.oncreate so that popover content generation is handled in a
separate function (createPopoverContent) so that we can inherit from
Field for more customized displays.

Add the EVRField component for displaying EVR telemetry fields with
additional EVR data provided to the EVRDefinition's message. This allows
the display of EVR messages formatted with telemetry data.

For example, you have an EVR field named EVRHistCode0 and
your H&S data includes 4 U8 fields that are used to store data
relevant to the EVR reported in EVRHistCode0. We'll call them EVRHistData0,
EVRHistData1, EVRHistData2, and EVRHistData3. You could specify this
relationship in your MOS GUI config with:

<ait-evrfield packet="AIT_1553_HS" name="EVRHistCode0">
   <ait-evr-data packet="AIT_1553_HS" name="EVRHistData0"></ait-evr-data>
   <ait-evr-data packet="AIT_1553_HS" name="EVRHistData1"></ait-evr-data>
   <ait-evr-data packet="AIT_1553_HS" name="EVRHistData2"></ait-evr-data>
   <ait-evr-data packet="AIT_1553_HS" name="EVRHistData3"></ait-evr-data>
</ait-evrfield>

This displays the EVR field the same way the standard Field
component. However the telemetry dictionary popover would contain the
EVR message formatted with the data in the 4 fields mentioned. If the
EVR field being displayed doesn't have a message attribute it falls back
to displaying the standard Field popover.