Issue #102 - Remove fsevents dependency and fix mithril-query version
Remove fsevents as a dev dependency since it is OS X specific. The
dependency that needs it should be able to handle installing it without
us explicitly listing it. This should avoid issues on non-OS X machines.

Fix mithril query to 2.2.0 instead of ^2.2.0. New changes (seamingly in
2.5.0, but not 100% sure) were causing tests to break because changes to
component attributes weren't being reflected outside of the mq call as
expected. A clean install with the previous version requirements
resulted in 5 broken tests for me. After fixing it to 2.2.0 all tests
work on a clean install.