Issue #75 - Add check for log and telem stream name attributes
When a message is received the GUI now checks to see if the topic
is within its log and telem stream name attributes, and if so,
processes the message accordingly. If these attributes do not exist
or if the received topic is in not in them, the GUI checks if the
topic contains either "telem_stream" or "log_stream" and processes
accordingly. If none of these conditions are met, an error is
thrown that a message with an unrecognized topic has been received.