Issue #110 - Made GUI port optionally configurable
Instead of defaulting to port 8080, the ait gui should now look at the
ait config and pull the defined gui.port number from there if its defined.
If it is not defined, then it will use the default of 8080. All of which
gets overwritten if the user provided a port number to the init method