Issue #156 - Update Python version in RTD config
Update the Python version in RTD config

Monkey patch gevent in the sphinx conf.py file to ensure that monkey
patching is done prior to any other imports. This is to avoid a known
issue where SSL is imported prior to patching and an infinite recursion
problem is encountered.

See AIT-Core/issues/253

Bump GUI Sphinx version to match Core