Issue #194 - Fix improperly initialized 'dntoeu' packet states
Packet diff handling when previously unseen packet was received failed
to properly initialized the `dntoeu` values for that packet. This meant
that UI clients would be unable to view the "human readable" versions of
fields in the GUI for any field that updated infrequently. This updates
the new packet type handling to properly update the `dntoeu` values as
well.

Resolve #194