Issue #198 - Address stagnant DN to EU value issue
Update the packet diffing code so a fields raw / DNtoEU value is
always included in the diff if a DNtoEU is defined for that field.
This ensures that fields with "complex" DNtoEU functions which take
multiple input values or rely on other means to determine a value will
always be updated. This does mean that we're performing some
"unnecessary" DNtoEU conversions but in practice I expect this to have
no noticeable impact on performance.

Resolve #198